requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-ckeditor==6.0.0
django-cors-headers==3.6.0
django-cuser==2017.3.16
django-filter==2.4.0
django-js-asset==1.2.2
django-modeltranslation==0.16.1
django-rest-auth==0.9.5
django-rest-authtoken==2.1.3
django-rest-swagger==2.2.0
django-searchable-select==1.5.0
django-tinymce==3.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.8
Pillow==8.1.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
pytz==2020.5
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
